+++ src/Symfony/Component/Cache/Adapter/PdoAdapter.php
@@ -12,10 +12,9 @@
 namespace Symfony\Component\Cache\Adapter;
 
 use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\DBALException;
 use Doctrine\DBAL\Driver\ServerInfoAwareConnection;
 use Doctrine\DBAL\DriverManager;
-use Doctrine\DBAL\Exception;
+use Doctrine\DBAL\Exception as DBALException;
 use Doctrine\DBAL\Exception\TableNotFoundException;
 use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Schema\Schema;
@@ -108,7 +107,6 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
      *
      * @throws \PDOException    When the table already exists
      * @throws DBALException    When the table already exists
-     * @throws Exception        When the table already exists
      * @throws \DomainException When an unsupported PDO driver is used
      */
     public function createTable()
@@ -121,7 +119,7 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
             $this->addTableToSchema($schema);
 
             foreach ($schema->toSql($conn->getDatabasePlatform()) as $sql) {
-                if (method_exists($conn, 'executeStatement')) {
+                if ($conn instanceof Connection && method_exists($conn, 'executeStatement')) {
                     $conn->executeStatement($sql);
                 } else {
                     $conn->exec($sql);
@@ -156,7 +154,7 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
                 throw new \DomainException(sprintf('Creating the cache table is currently not implemented for PDO driver "%s".', $this->driver));
         }
 
-        if (method_exists($conn, 'executeStatement')) {
+        if ($conn instanceof Connection && method_exists($conn, 'executeStatement')) {
             $conn->executeStatement($sql);
         } else {
             $conn->exec($sql);
@@ -305,7 +303,7 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
         }
 
         try {
-            if (method_exists($conn, 'executeStatement')) {
+            if ($conn instanceof Connection && method_exists($conn, 'executeStatement')) {
                 $conn->executeStatement($sql);
             } else {
                 $conn->exec($sql);
@@ -435,7 +433,7 @@ class PdoAdapter extends AbstractAdapter implements PruneableInterface
             if (null === $driver && !(\is_object($result) ? $result->rowCount() : $stmt->rowCount())) {
                 try {
                     $insertStmt->execute();
-                } catch (DBALException | Exception $e) {
+                } catch (DBALException $e) {
                 } catch (\PDOException $e) {
                     // A concurrent write won, let it be
                 }