+++ src/Symfony/Component/DependencyInjection/ContainerBuilder.php
@@ -1211,10 +1211,10 @@ class ContainerBuilder extends Container implements TaggedContainerInterface
     {
         $tags = [];
         foreach ($this->getDefinitions() as $id => $definition) {
-            $tags = array_merge(array_keys($definition->getTags()), $tags);
+            $tags[] = array_keys($definition->getTags());
         }
 
-        return array_unique($tags);
+        return array_unique(array_merge([], ...$tags));
     }
 
     /**