+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -6,6 +6,10 @@ CHANGELOG
 
  * Remove `DoctrineTestHelper` and `TestRepositoryFactory`
 
+5.4
+---
+ * Add a middleware to log when transaction has been left open `DoctrineOpenTransactionLoggerMiddleware`
+
 5.3
 ---